feat: support negative dim for index.select
